unit test string collection
